v1.4.6 Insere duas queries pra evitar JOIN FETCH e warning de ineficiência; e também para excluir LIMIT da JPQL que é fora do padrão e só funciona em alguns casos

+++ src/main/java/io/github/marciocg/clientes/ClientesResource.java
@@ -25,7 +25,7 @@ public class ClientesResource {
     @Path("/{id}/extrato")
 
     public ExtratoResponseDTO extrato(@PathParam("id") Integer id) {
-        Saldo saldo = Saldo.getSaldoWithUltimasTransacoesById(BigInteger.valueOf(id));
+        Saldo saldo = Saldo.getSaldoWithUltimasTransacoesById(id);
         if (saldo == null) {
             saldo = Saldo.getSaldoById(id.intValue());
         }
